+++ src/main/types.ts
@@ -343,6 +343,14 @@ type AppState = {
   preferredViewpoint: string;
 };
 
+type AdvancedLoggingStatus = {
+  retail: boolean;
+  classic: boolean;
+  era: boolean;
+  retailPtr: boolean;
+  classicPtr: boolean;
+};
+
 type CloudState = {
   uploadProgress: number;
   downloadProgress: number;
@@ -691,4 +699,5 @@ export {
   SoundAlerts,
   CloudState,
   ActivityStatus,
+  AdvancedLoggingStatus,
 };